refactoring of ANIMATOR-PLAYBACK.test.ts
