+++ src/test/ANIMATOR_PLAYBACK.test.ts
@@ -1,48 +1,80 @@
 import { CARTA } from "carta-protobuf";
-import { Client } from "./CLIENT";
+import { Client, AckStream } from "./CLIENT";
 import config from "./config.json";
 import * as Long from "long";
-let testServerUrl = config.serverURL;
-let testSubdirectory = config.path.QA;
-let connectTimeout = config.timeout.connection;
-let playTimeout = config.timeout.playImages;
+
+let testServerUrl: string = config.serverURL;
+let testSubdirectory: string = config.path.QA;
+let connectTimeout: number = config.timeout.connection;
+let openFileTimeout: number = config.timeout.openFile;
+let playImageTimeout: number = 10000;//config.timeout.playImages;
+let sleepTimeout: number = config.timeout.sleep
+let playAnimatorTimeout = config.timeout.playAnimator
+
 interface AssertItem {
     register: CARTA.IRegisterViewer;
-    fileOpens: CARTA.IOpenFile;
+    filelist: CARTA.IFileListRequest;
+    fileOpen: CARTA.IOpenFile;
+    addTilesReq: CARTA.IAddRequiredTiles[];
+    setCursor: CARTA.ISetCursor;
+    setSpatialReq: CARTA.ISetSpatialRequirements;
     startAnimation: CARTA.IStartAnimation[];
-    stopAnimation: CARTA.IStopAnimation;
-    animationFlowControl: CARTA.IAnimationFlowControl;
+    stopAnimation: CARTA.IStopAnimation[];
+    animationFlowControl: CARTA.IAnimationFlowControl[];
+    setImageChannel: CARTA.ISetImageChannels[];
     reverseAnimation: CARTA.IStartAnimation[];
     blinkAnimation: CARTA.IStartAnimation;
 };
+
 let assertItem: AssertItem = {
     register: {
         sessionId: 0,
-        apiKey: "",
         clientFeatureFlags: 5,
     },
-    fileOpens:
-    {
+    filelist: { directory: testSubdirectory },
+    fileOpen: {
         directory: testSubdirectory,
         file: "M17_SWex.image",
-        fileId: 0,
         hdu: "",
+        fileId: 0,
         renderMode: CARTA.RenderMode.RASTER,
+    },
+    addTilesReq: [
+        {
+            fileId: 0,
+            compressionQuality: 11,
+            compressionType: CARTA.CompressionType.ZFP,
+            tiles: [33558529, 33558528, 33562625, 33554433, 33562624, 33558530, 33554432, 33562626, 33554434, 33566721, 33566720, 33566722],
+        },
+        {
+            fileId: 0,
+            compressionQuality: 9,
+            compressionType: CARTA.CompressionType.ZFP,
+            tiles: [33554432, 33558528, 33562624, 33566720, 33554433, 33558529, 33562625, 33566721, 33554434, 33558530, 33562626, 33566722],
+        },
+    ],
+    setCursor: {
+        fileId: 0,
+        point: { x: 1, y: 1 },
+    },
+    setSpatialReq: {
+        fileId: 0,
+        regionId: 0,
+        spatialProfiles: ["x", "y"]
     },
     startAnimation:
         [
             {
                 fileId: 0,
-                startFrame: { channel: 10, stokes: 0 },
-                firstFrame: { channel: 10, stokes: 0 },
-                lastFrame: { channel: 20, stokes: 0 },
+                startFrame: { channel: 1, stokes: 0 },
+                firstFrame: { channel: 0, stokes: 0 },
+                lastFrame: { channel: 24, stokes: 0 },
                 deltaFrame: { channel: 1, stokes: 0 },
-                imageView: {
+                requiredTiles: {
                     fileId: 0,
-                    imageBounds: { xMax: 640, yMax: 800 },
-                    mip: 4,
+                    tiles: [33554432, 33558528, 33562624, 33566720, 33554433, 33558529, 33562625, 33566721, 33554434, 33558530, 33562626, 33566722],
                     compressionType: CARTA.CompressionType.ZFP,
-                    compressionQuality: 16,
+                    compressionQuality: 9,
                 },
             },
             {
@@ -51,25 +83,109 @@ let assertItem: AssertItem = {
                 firstFrame: { channel: 9, stokes: 0 },
                 lastFrame: { channel: 19, stokes: 0 },
                 deltaFrame: { channel: -1, stokes: 0 },
-                imageView: {
+                requiredTiles: {
                     fileId: 0,
-                    imageBounds: { xMax: 640, yMax: 800 },
-                    mip: 4,
+                    tiles: [33554432, 33558528, 33562624, 33566720, 33554433, 33558529, 33562625, 33566721, 33554434, 33558530, 33562626, 33566722],
                     compressionType: CARTA.CompressionType.ZFP,
-                    compressionQuality: 16,
+                    compressionQuality: 9,
+                },
+            },
+            {
+                fileId: 0,
+                startFrame: { channel: 9, stokes: 0 },
+                firstFrame: { channel: 9, stokes: 0 },
+                lastFrame: { channel: 19, stokes: 0 },
+                deltaFrame: { channel: 1, stokes: 0 },
+                requiredTiles: {
+                    fileId: 0,
+                    tiles: [33554432, 33558528, 33562624, 33566720, 33554433, 33558529, 33562625, 33566721, 33554434, 33558530, 33562626, 33566722],
+                    compressionType: CARTA.CompressionType.ZFP,
+                    compressionQuality: 9,
                 },
             },
         ],
     stopAnimation:
-    {
-        fileId: 0,
-        endFrame: { channel: 18, stokes: 0 },
-    },
+        [
+            {
+                fileId: 0,
+                endFrame: { channel: 10, stokes: 0 },
+            },
+            {
+                fileId: 0,
+                endFrame: { channel: 19, stokes: 0 },
+            },
+            {
+                fileId: 0,
+                endFrame: { channel: 9, stokes: 0 },
+            },
+        ],
     animationFlowControl:
-    {
-        fileId: 0,
-        animationId: 0,
-    },
+        [
+            {
+                fileId: 0,
+                animationId: 1,
+            },
+            {
+                fileId: 0,
+                animationId: 2,
+            },
+            {
+                fileId: 0,
+                animationId: 3,
+            },
+            {
+                fileId: 0,
+                animationId: 4,
+            },
+            {
+                fileId: 0,
+                animationId: 5,
+            },
+            {
+                fileId: 0,
+                animationId: 6,
+            },
+            {
+                fileId: 0,
+                animationId: 7,
+            },
+        ],
+    setImageChannel:
+        [
+            {
+                fileId: 0,
+                channel: 10,
+                stokes: 0,
+                requiredTiles: {
+                    fileId: 0,
+                    tiles: [33558529, 33558528, 33562625, 33554433, 33562624, 33558530, 33554432, 33562626, 33554434, 33566721, 33566720, 33566722],
+                    compressionType: CARTA.CompressionType.ZFP,
+                    compressionQuality: 11,
+                },
+            },
+            {
+                fileId: 0,
+                channel: 19,
+                stokes: 0,
+                requiredTiles: {
+                    fileId: 0,
+                    tiles: [33558529, 33558528, 33562625, 33554433, 33562624, 33558530, 33554432, 33562626, 33554434, 33566721, 33566720, 33566722],
+                    compressionType: CARTA.CompressionType.ZFP,
+                    compressionQuality: 11,
+                },
+            },
+            {
+                fileId: 0,
+                channel: 9,
+                stokes: 0,
+                requiredTiles: {
+                    fileId: 0,
+                    tiles: [33558529, 33558528, 33562625, 33554433, 33562624, 33558530, 33554432, 33562626, 33554434, 33566721, 33566720, 33566722],
+                    compressionType: CARTA.CompressionType.ZFP,
+                    compressionQuality: 11,
+                },
+            },
+        ],
     reverseAnimation:
         [
             {
@@ -78,12 +194,11 @@ let assertItem: AssertItem = {
                 firstFrame: { channel: 10, stokes: 0 },
                 lastFrame: { channel: 20, stokes: 0 },
                 deltaFrame: { channel: 1, stokes: 0 },
-                imageView: {
+                requiredTiles: {
                     fileId: 0,
-                    imageBounds: { xMax: 640, yMax: 800 },
-                    mip: 4,
+                    tiles: [33554432, 33558528, 33562624, 33566720, 33554433, 33558529, 33562625, 33566721, 33554434, 33558530, 33562626, 33566722],
                     compressionType: CARTA.CompressionType.ZFP,
-                    compressionQuality: 16,
+                    compressionQuality: 9,
                 },
                 reverse: true,
             },
@@ -93,12 +208,11 @@ let assertItem: AssertItem = {
                 firstFrame: { channel: 10, stokes: 0 },
                 lastFrame: { channel: 20, stokes: 0 },
                 deltaFrame: { channel: -1, stokes: 0 },
-                imageView: {
+                requiredTiles: {
                     fileId: 0,
-                    imageBounds: { xMax: 640, yMax: 800 },
-                    mip: 4,
+                    tiles: [33554432, 33558528, 33562624, 33566720, 33554433, 33558529, 33562625, 33566721, 33554434, 33558530, 33562626, 33566722],
                     compressionType: CARTA.CompressionType.ZFP,
-                    compressionQuality: 16,
+                    compressionQuality: 9,
                 },
             },
         ],
@@ -109,18 +223,18 @@ let assertItem: AssertItem = {
         firstFrame: { channel: 3, stokes: 0 },
         lastFrame: { channel: 10, stokes: 0 },
         deltaFrame: { channel: 7, stokes: 0 },
-        imageView: {
+        requiredTiles: {
             fileId: 0,
-            imageBounds: { xMax: 640, yMax: 800 },
-            mip: 4,
+            tiles: [33554432, 33558528, 33562624, 33566720, 33554433, 33558529, 33562625, 33566721, 33554434, 33558530, 33562626, 33566722],
             compressionType: CARTA.CompressionType.ZFP,
-            compressionQuality: 16,
+            compressionQuality: 9,
         },
         looping: true,
     },
 };
 
 describe("ANIMATOR_PLAYBACK test: Testing animation playback", () => {
+
     let Connection: Client;
     beforeAll(async () => {
         Connection = new Client(testServerUrl);
@@ -129,256 +243,312 @@ describe("ANIMATOR_PLAYBACK test: Testing animation playback", () => {
         await Connection.receive(CARTA.RegisterViewerAck);
     }, connectTimeout);
 
-    describe(`Go to "${testSubdirectory}" folder and open images`, () => {
+    test(`(Step 0) Connection open? | `, () => {
+        expect(Connection.connection.readyState).toBe(WebSocket.OPEN);
+    });
 
+    function sleep(ms) {
+        return new Promise(resolve => setTimeout(resolve, ms)).then(() => { console.log('sleep!') });
+    }
+
+    describe(`Go to "${assertItem.filelist.directory}" folder`, () => {
         beforeAll(async () => {
             await Connection.send(CARTA.CloseFile, { fileId: -1 });
-            await Connection.send(CARTA.OpenFile, assertItem.fileOpens);
-            await Connection.receiveAny();
-            await Connection.receiveAny(); // OpenFileAck | RegionHistogramData
+        }, connectTimeout);
+
+        describe(`(Step 1) Initialization: the open image`, () => {
+            test(`OPEN_FILE_ACK and REGION_HISTOGRAM_DATA should arrive within ${openFileTimeout} ms`, async () => {
+                await Connection.send(CARTA.CloseFile, { fileId: 0 });
+                await Connection.send(CARTA.OpenFile, assertItem.fileOpen);
+                await Connection.receiveAny()
+                await Connection.receiveAny() // OpenFileAck | RegionHistogramData
+            }, openFileTimeout);
+
+            let ack: AckStream;
+            test(`return RASTER_TILE_DATA(Stream) and check total length `, async () => {
+                await Connection.send(CARTA.AddRequiredTiles, assertItem.addTilesReq[0]);
+                await Connection.send(CARTA.SetCursor, assertItem.setCursor);
+                await Connection.send(CARTA.SetSpatialRequirements, assertItem.setSpatialReq);
+
+                ack = await Connection.stream(15) as AckStream;
+                // console.log(ack); // RasterTileData * 12 + SpatialProfileData * 1 + RasterTileSync *2 (start & end)
+                expect(ack.RasterTileData.length).toBe(assertItem.addTilesReq[0].tiles.length);
+            }, playImageTimeout);
+
         });
 
-        describe(`Play all images forwardly with looping`, () => {
-            let RasterImageData: CARTA.RasterImageData[] = [];
+        describe(`(Step 2):Play some channels forwardly`, () => {
+            let AnimateStreamData: AckStream[] = [];
             let sequence: number[] = [];
-            test(`Image should return one after one`, async () => {
+            test(`Image should return one after one and the last channel is correct:`, async () => {
                 await Connection.send(CARTA.StartAnimation, {
                     ...assertItem.startAnimation[0],
                     looping: true,
+                    reverse: false,
+                    frameRate: 5,
                 });
-                await Connection.receive(CARTA.StartAnimationAck);
-                for (let i = 0; i < 2 * Math.abs(assertItem.startAnimation[0].lastFrame.channel - assertItem.startAnimation[0].firstFrame.channel) + 1; i++) {
-                    RasterImageData.push(await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData);
+                await Connection.send(CARTA.AddRequiredTiles, assertItem.addTilesReq[1]);
+                let SAAck = await Connection.receive(CARTA.StartAnimationAck);
+                expect(SAAck.success).toBe(true);
+
+                for (let i = 0; i < assertItem.stopAnimation[0].endFrame.channel; i++) {
+                    AnimateStreamData.push(await Connection.stream(16) as AckStream);
                     await Connection.send(CARTA.AnimationFlowControl,
                         {
-                            ...assertItem.animationFlowControl,
+                            ...assertItem.animationFlowControl[0],
                             receivedFrame: {
-                                channel: RasterImageData[i].channel,
+                                channel: AnimateStreamData[i].RasterTileData[0].channel,
                                 stokes: 0
                             },
                             timestamp: Long.fromNumber(Date.now()),
                         }
                     );
-                    sequence.push(RasterImageData[i].channel);
-                }
-                await Connection.send(CARTA.StopAnimation, CARTA.StopAnimation);
-                let lastRasterImageData = await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData;
-                sequence.push(lastRasterImageData.channel);
-            }, playTimeout);
+                    // console.log(AnimateStreamData[i].RasterTileData[0].channel) // In principle, each channel should have RasterTileSync *2 (start & end)
+                    sequence.push(AnimateStreamData[i].RasterTileData[0].channel);
+                };
+
+                await Connection.send(CARTA.StopAnimation, assertItem.stopAnimation[0]);
+                await Connection.send(CARTA.SetImageChannels, assertItem.setImageChannel[0])
+                let lastRasterImageData = await Connection.stream(14) as AckStream;
+                // console.log(lastRasterImageData); // RasterTileData * 12 + RasterTileSync *2 (start & end)
+                // console.log(AnimateStreamData); // RasterTileData * 12 + SpatialProfileData * 1 + RegionHistogramData * 1 + RasterTileSync *2 (start & end)
+                // console.log(sequence); // show looping sequence
+                expect(sequence[sequence.length - 1]).toEqual(assertItem.stopAnimation[0].endFrame.channel);
+
+            }, playAnimatorTimeout)
 
             test(`Received image channels should be in sequence`, async () => {
-                console.log(`Sequent channel index: ${sequence}`);
-                let i = 0;
-                RasterImageData.map((imageData, index) => {
-                    let j = i++ % Math.abs(1 + assertItem.startAnimation[0].lastFrame.channel - assertItem.startAnimation[0].firstFrame.channel) + assertItem.startAnimation[0].firstFrame.channel;
-                    expect(imageData.channel).toEqual(j);
-                });
-            });
-
-            test(`Received image size should be ${JSON.stringify(assertItem.startAnimation[0].imageView.imageBounds)}`, async () => {
-                RasterImageData.map((imageData, index) => {
-                    expect(imageData.imageBounds).toEqual(assertItem.startAnimation[0].imageView.imageBounds);
+                console.warn(`(Step 2) Sequent channel index: ${sequence}`);
+                AnimateStreamData.map((imageData, index) => {
+                    let j = (index + assertItem.startAnimation[0].startFrame.channel + assertItem.startAnimation[0].deltaFrame.channel) - 1.;
+                    // let i = 3 * Math.abs(assertItem.stopAnimation[0].endFrame.channel - assertItem.startAnimation[0].firstFrame.channel) + 5;
+                    // let j = i-- % Math.abs(1 + assertItem.startAnimation[0].lastFrame.channel - assertItem.startAnimation[0].firstFrame.channel) + assertItem.startAnimation[0].firstFrame.channel;
+                    expect(sequence[index]).toEqual(j);
                 });
             });
         });
 
-        describe(`Play all images backwardly with looping`, () => {
-            let RasterImageData: CARTA.RasterImageData[] = [];
+        describe(`(Step 3) Play some channels backwardly with looping`, () => {
+            let AnimateStreamData: AckStream[] = [];
             let sequence: number[] = [];
             test(`Image should return one after one`, async () => {
+                await sleep(sleepTimeout);
                 await Connection.send(CARTA.StartAnimation, {
                     ...assertItem.startAnimation[1],
                     looping: true,
+                    reverse: false,
+                    frameRate: 5,
                 });
+                await Connection.send(CARTA.AddRequiredTiles, assertItem.addTilesReq[1]);
                 await Connection.receive(CARTA.StartAnimationAck);
-                for (let i = 0; i < 2 * Math.abs(assertItem.startAnimation[1].lastFrame.channel - assertItem.startAnimation[1].firstFrame.channel) + 1; i++) {
-                    RasterImageData.push(await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData);
+                for (let i = 0; i < 2 * Math.abs(assertItem.startAnimation[1].lastFrame.channel - assertItem.startAnimation[1].firstFrame.channel + 1); i++) {
+                    AnimateStreamData.push(await Connection.stream(16) as AckStream);
                     await Connection.send(CARTA.AnimationFlowControl,
                         {
-                            ...assertItem.animationFlowControl,
+                            ...assertItem.animationFlowControl[1],
                             receivedFrame: {
-                                channel: RasterImageData[i].channel,
-                                stokes: 0
+                                channel: AnimateStreamData[i].RasterTileData[0].channel,
+                                stokes: 0,
                             },
                             timestamp: Long.fromNumber(Date.now()),
                         }
                     );
-                    sequence.push(RasterImageData[i].channel);
-                }
-                await Connection.send(CARTA.StopAnimation, CARTA.StopAnimation);
-                let lastRasterImageData = await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData;
-                sequence.push(lastRasterImageData.channel);
-            }, playTimeout);
+                    sequence.push(AnimateStreamData[i].RasterTileData[0].channel);
+                };
+                // console.log(AnimateStreamData); // RasterTileData * 12 + SpatialProfileData * 1 + RegionHistogramData * 1 + RasterTileSync *2 (start & end)
+
+                await Connection.send(CARTA.StopAnimation, assertItem.stopAnimation[1]);
+                await Connection.send(CARTA.SetImageChannels, assertItem.setImageChannel[1]);
+                let lastRasterImageData = await Connection.stream(16) as AckStream;
+                // console.log(lastRasterImageData); // RasterTileData * 12 + SpatialProfileData * 1 + RegionHistogramData * 1 + RasterTileSync *2 (start & end)
+                sequence.push(lastRasterImageData.RasterTileData[0].channel);
+                // console.log(sequence);
+
+                expect(sequence[sequence.length - 1]).toEqual(assertItem.stopAnimation[1].endFrame.channel);
+            }, playAnimatorTimeout);
 
             test(`Received image channels should be in sequence`, async () => {
-                console.log(`Sequent channel index: ${sequence}`);
+                console.warn(`(Step 3) Sequent channel index: ${sequence}`);
                 let i = 2 * Math.abs(assertItem.startAnimation[1].lastFrame.channel - assertItem.startAnimation[1].firstFrame.channel) + 1;
-                RasterImageData.map((imageData, index) => {
+                AnimateStreamData.map((imageData, index) => {
                     let j = i-- % Math.abs(1 + assertItem.startAnimation[1].lastFrame.channel - assertItem.startAnimation[1].firstFrame.channel) + assertItem.startAnimation[1].firstFrame.channel;
-                    expect(imageData.channel).toEqual(j);
+                    expect(imageData.RasterTileData[0].channel).toEqual(j);
                 });
             });
 
-            test(`Received image size should be ${JSON.stringify(assertItem.startAnimation[1].imageView.imageBounds)}`, async () => {
-                RasterImageData.map((imageData, index) => {
-                    expect(imageData.imageBounds).toEqual(assertItem.startAnimation[1].imageView.imageBounds);
-                });
-            });
         });
 
-        describe(`Play some images until stop`, () => {
-            let RasterImageData: CARTA.RasterImageData[] = [];
-            test(`Image should return one after one`, async () => {
-                await Connection.send(CARTA.StartAnimation, assertItem.startAnimation[0]);
-                await Connection.receive(CARTA.StartAnimationAck);
-                for (let i = 0; i < Math.abs(assertItem.stopAnimation.endFrame.channel - assertItem.startAnimation[0].firstFrame.channel); i++) {
-                    RasterImageData.push(await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData);
-                    await Connection.send(CARTA.AnimationFlowControl,
-                        {
-                            ...assertItem.animationFlowControl,
-                            receivedFrame: {
-                                channel: RasterImageData[i].channel,
-                                stokes: 0
-                            },
-                            timestamp: Long.fromNumber(Date.now()),
-                        }
-                    );
-                }
-            }, playTimeout);
-
-            test(`Last image on channel${JSON.stringify(assertItem.stopAnimation.endFrame)} should receive after stop`, async () => {
-                await Connection.send(CARTA.StopAnimation, assertItem.stopAnimation);
-                let RasterImageData = await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData;
-                expect(RasterImageData.channel).toEqual(assertItem.stopAnimation.endFrame.channel);
-            });
-        });
-
-        describe(`Play images round-trip`, () => {
-            let RasterImageData: CARTA.RasterImageData[] = [];
+        describe(`(Step 4 )Play some channels forwardly with looping until stop`, () => {
+            let AnimateStreamData: AckStream[] = [];
+            let lastRasterImageData: AckStream;
             let sequence: number[] = [];
             test(`Image should return one after one`, async () => {
-                await Connection.send(CARTA.StartAnimation,
-                    {
-                        ...assertItem.startAnimation[0],
-                        reverse: true,
-                    }
-                );
+                await sleep(sleepTimeout);
+                await Connection.send(CARTA.StartAnimation, {
+                    ...assertItem.startAnimation[2],
+                    looping: true,
+                    reverse: false,
+                    frameRate: 5,
+                });
+                await Connection.send(CARTA.AddRequiredTiles, assertItem.addTilesReq[1]);
                 await Connection.receive(CARTA.StartAnimationAck);
-                for (let i = 0; i < 3 * Math.abs(assertItem.startAnimation[0].lastFrame.channel - assertItem.startAnimation[0].firstFrame.channel); i++) {
-                    RasterImageData.push(await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData);
+
+                for (let i = 0; i < 2 * Math.abs(assertItem.startAnimation[2].lastFrame.channel - assertItem.startAnimation[2].firstFrame.channel + 1); i++) {
+                    // console.log(i);
+                    AnimateStreamData.push(await Connection.stream(16) as AckStream);
                     await Connection.send(CARTA.AnimationFlowControl,
                         {
-                            ...assertItem.animationFlowControl,
+                            ...assertItem.animationFlowControl[2],
                             receivedFrame: {
-                                channel: RasterImageData[i].channel,
+                                channel: AnimateStreamData[i].RasterTileData[0].channel,
                                 stokes: 0
                             },
                             timestamp: Long.fromNumber(Date.now()),
                         }
                     );
-                    sequence.push(RasterImageData[i].channel);
-                }
-                await Connection.send(CARTA.StopAnimation, CARTA.StopAnimation);
-                let lastRasterImageData = await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData;
-                sequence.push(lastRasterImageData.channel);
-            }, playTimeout);
+                    sequence.push(AnimateStreamData[i].RasterTileData[0].channel);
+                    // console.log(AnimateStreamData[i].RasterTileData);
+                };
+                await Connection.send(CARTA.StopAnimation, assertItem.stopAnimation[2]);
+                await Connection.send(CARTA.SetImageChannels, assertItem.setImageChannel[2]);
 
-            test(`Received image channels should be in sequence and then reverse`, async () => {
-                console.log(`Channel index in roundtrip: ${sequence}`);
-                let previous: number = assertItem.startAnimation[0].startFrame.channel - 1;
-                RasterImageData.map((imageData, index) => {
-                    expect(Math.abs(imageData.channel - previous)).toEqual(1);
-                    previous = imageData.channel;
-                });
+                lastRasterImageData = await Connection.stream(16) as AckStream;
+                sequence.push(lastRasterImageData.RasterTileData[0].channel);
+                // console.log(AnimateStreamData);
+                // console.log(lastRasterImageData)
+
+            }, playAnimatorTimeout);
+
+            test(`Last channel should be received after stop and should be ${JSON.stringify(assertItem.stopAnimation[2].endFrame.channel)}`, async () => {
+                console.warn(`(Step 4) Sequent channel index: ${sequence}`);
+                console.warn('(Step 4) Last Raster Tile channel:', lastRasterImageData.RasterTileData[0].channel);
+                expect(lastRasterImageData.RasterTileData[0].channel).toEqual(assertItem.stopAnimation[2].endFrame.channel);
             });
+        });
 
-            test(`Received image size should be ${JSON.stringify(assertItem.startAnimation[0].imageView.imageBounds)}`, async () => {
-                RasterImageData.map((imageData, index) => {
-                    expect(imageData.imageBounds).toEqual(assertItem.startAnimation[0].imageView.imageBounds);
+        describe(`(Step 5) Play images round-trip`, () => {
+            let AnimateStreamData: AckStream[] = [];
+            let sequence: number[] = [];
+            test(`Image should return one after one`, async () => {
+                await sleep(sleepTimeout);
+                await Connection.send(CARTA.StartAnimation, {
+                    ...assertItem.startAnimation[0],
+                    reverse: true,
                 });
+                await Connection.send(CARTA.AddRequiredTiles, assertItem.addTilesReq[1]);
+                await Connection.receive(CARTA.StartAnimationAck);
+
+                for (let i = 0; i < 3 * Math.abs(assertItem.startAnimation[0].lastFrame.channel - assertItem.startAnimation[0].firstFrame.channel); i++) {
+                    // console.log(i);
+                    AnimateStreamData.push(await Connection.stream(16) as AckStream);
+                    await Connection.send(CARTA.AnimationFlowControl,
+                        {
+                            ...assertItem.animationFlowControl[3],
+                            receivedFrame: {
+                                channel: AnimateStreamData[i].RasterTileData[0].channel,
+                                stokes: 0,
+                            },
+                            timestamp: Long.fromNumber(Date.now()),
+                        }
+                    );
+                    sequence.push(AnimateStreamData[i].RasterTileData[0].channel);
+                };
+                await Connection.send(CARTA.StopAnimation, assertItem.stopAnimation[2]);
+                await Connection.send(CARTA.SetImageChannels, assertItem.setImageChannel[2])
+                let lastRasterImageData = await Connection.stream(16) as AckStream;
+                // console.log(lastRasterImageData)
+                // console.log(sequence);
+
+            }, playAnimatorTimeout);
+
+            test(`Received image channels should be in sequence and then reverse:`, async () => {
+                console.warn(`(Step 5) Channel index in roundtrip: ${sequence}`);
+                let previous: number = assertItem.startAnimation[0].lastFrame.channel;
+                // console.log(previous);
+                expect(Math.abs(sequence[sequence.length - 1] - previous)).toEqual(0);
             });
         });
 
         assertItem.reverseAnimation.map((animation, index) => {
-            describe(`Play all images backwardly using method${index + 1}`, () => {
-                let RasterImageData: CARTA.RasterImageData[] = [];
+            describe(`(Step 6) Play all images backwardly using method${index + 1}`, () => {
+                let AnimateStreamData: AckStream[] = [];
                 let sequence: number[] = [];
                 test(`Image should return one after one`, async () => {
+                    await sleep(sleepTimeout);
                     await Connection.send(CARTA.StartAnimation, animation);
                     await Connection.receive(CARTA.StartAnimationAck);
                     for (let i = 0; i < Math.abs(animation.lastFrame.channel - animation.firstFrame.channel); i++) {
-                        RasterImageData.push(await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData);
+                        AnimateStreamData.push(await Connection.stream(16) as AckStream);
                         await Connection.send(CARTA.AnimationFlowControl,
                             {
-                                ...assertItem.animationFlowControl,
+                                ...assertItem.animationFlowControl[4 + index],
                                 receivedFrame: {
-                                    channel: RasterImageData[i].channel,
+                                    channel: AnimateStreamData[i].RasterTileData[0].channel,
                                     stokes: 0
                                 },
                                 timestamp: Long.fromNumber(Date.now()),
                             }
                         );
-                        sequence.push(RasterImageData[i].channel);
+                        sequence.push(AnimateStreamData[i].RasterTileData[0].channel);
                     }
-                    await Connection.send(CARTA.StopAnimation, CARTA.StopAnimation);
-                    let lastRasterImageData = await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData;
-                    sequence.push(lastRasterImageData.channel);
-                }, playTimeout);
+                    await Connection.send(CARTA.StopAnimation, assertItem.stopAnimation[0]);
+                    await Connection.send(CARTA.SetImageChannels, assertItem.setImageChannel[0])
+                    let lastRasterImageData = await Connection.stream(16) as AckStream;
+                    // console.log(lastRasterImageData);
+                    // console.log(sequence);
+                }, playAnimatorTimeout);
 
                 test(`Received image channels should be in sequence`, async () => {
-                    console.log(`Backward channel index: ${sequence}`);
-                    RasterImageData.map((imageData, index) => {
-                        expect(imageData.channel).toEqual(animation.startFrame.channel - index);
-                    });
-                });
-
-                test(`Received image size should be ${JSON.stringify(animation.imageView.imageBounds)}`, async () => {
-                    RasterImageData.map((imageData, index) => {
-                        expect(imageData.imageBounds).toEqual(animation.imageView.imageBounds);
+                    console.warn(`(Step 6) Backward channel index with method${index + 1}: ${sequence}`);
+                    AnimateStreamData.map((imageData, index) => {
+                        expect(sequence[index]).toEqual(animation.startFrame.channel - index);
                     });
                 });
             });
         });
 
-        describe(`Blink images bewteen ${assertItem.blinkAnimation.firstFrame.channel} and ${assertItem.blinkAnimation.lastFrame.channel}`, () => {
-            let RasterImageData: CARTA.RasterImageData[] = [];
+        describe(`(Step 7) Blink images between ${assertItem.blinkAnimation.firstFrame.channel} and ${assertItem.blinkAnimation.lastFrame.channel}`, () => {
+            let AnimateStreamData: AckStream[] = [];
             let sequence: number[] = [];
             test(`Image should return one after one`, async () => {
+                await sleep(sleepTimeout);
                 await Connection.send(CARTA.StartAnimation, assertItem.blinkAnimation);
                 await Connection.receive(CARTA.StartAnimationAck);
                 for (let i = 0; i < 11; i++) {
-                    RasterImageData.push(await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData);
+                    AnimateStreamData.push(await Connection.stream(16) as AckStream);
                     await Connection.send(CARTA.AnimationFlowControl,
                         {
-                            ...assertItem.animationFlowControl,
+                            ...assertItem.animationFlowControl[6],
                             receivedFrame: {
-                                channel: RasterImageData[i].channel,
+                                channel: AnimateStreamData[i].RasterTileData[0].channel,
                                 stokes: 0
                             },
                             timestamp: Long.fromNumber(Date.now()),
                         }
                     );
-                    sequence.push(RasterImageData[i].channel);
+                    sequence.push(AnimateStreamData[i].RasterTileData[0].channel);
                 }
-                await Connection.send(CARTA.StopAnimation, CARTA.StopAnimation);
-                let lastRasterImageData = await Connection.receive(CARTA.RasterImageData) as CARTA.RasterImageData;
-                sequence.push(lastRasterImageData.channel);
-            }, playTimeout);
+                await Connection.send(CARTA.StopAnimation, assertItem.startAnimation[0]);
+                await Connection.send(CARTA.SetImageChannels, assertItem.setImageChannel[0])
+                let lastRasterImageData = await Connection.stream(16) as AckStream;
+                // sequence.push(lastRasterImageData.channel);
+                // console.log(lastRasterImageData);
+                // console.log(sequence);
+            }, playAnimatorTimeout);
 
             test(`Received image channels should be in sequence`, async () => {
-                console.log(`Blink channel index: ${sequence}`);
-                RasterImageData.map((imageData, index) => {
-                    expect(imageData.channel === assertItem.blinkAnimation.firstFrame.channel || imageData.channel === assertItem.blinkAnimation.lastFrame.channel).toBe(true);
-                });
-            });
-
-            test(`Received image size should be ${JSON.stringify(assertItem.blinkAnimation.imageView.imageBounds)}`, async () => {
-                RasterImageData.map((imageData, index) => {
-                    expect(imageData.imageBounds).toEqual(assertItem.blinkAnimation.imageView.imageBounds);
+                console.warn(`(Step 7) Blink channel index: ${sequence}`);
+                AnimateStreamData.map((imageData, index) => {
+                    expect(imageData.RasterTileData[0].channel === assertItem.blinkAnimation.firstFrame.channel || imageData.RasterTileData[0].channel === assertItem.blinkAnimation.lastFrame.channel).toBe(true);
                 });
             });
         });
     });
 
-    afterAll(async () => await Connection.close());
+    test(`close file`, async () => {
+        await Connection.send(CARTA.CloseFile, { fileId: 0 });
+    }, connectTimeout);
+
+    afterAll(() => Connection.close());
+
 });
